fix(vault,engine): clear all clippy::all -D warnings violations for CI
Ubuntu CI runs `clippy --workspace --all-targets --all-features -D
warnings`; the prior commits introduced 12 violations across the
crypto + cloud + items + router + dynamic + sqlite + sigv4 + audit
surfaces + the engine_smoke test. Each one fixed:

- crypto/kek_store.rs — type-complexity on the load_active_* tuples
  (PG + SQLite). Extracted PgKekRow + SqliteKekRow type aliases.
- cloud/sigv4.rs — map(|(k,v)| (k,v)) → direct .collect();
  inputs.clone() on a Copy struct → just inputs (test only used it
  to call sign() twice, the second call moves a fresh copy).
- crypto/sealing.rs — `match { Ok(bad) => …, Err(_) => {} }` →
  `if let Ok(bad) = …` (single-pattern match).
- items.rs — create_item has 8 args mirroring the Item shape
  exactly; bundling them into a struct adds ceremony without
  clarity. Suppress `clippy::too_many_arguments` at the trait
  method with a documenting `#[allow]`.
- router.rs — Result<(), Response> trips clippy::result_large_err
  because axum::response::Response is hundreds of bytes; every
  call site propagates the response verbatim, so boxing the Err
  would force a deref at every site for no semantic gain.
  Suppress at the function with a documenting `#[allow]`.
- dynamic.rs — `ttl_secs.max(900) as u64` cast u64→u64; drop the
  cast.
- store/sqlite.rs — `repeat("?").take(n).collect()` →
  `repeat_n("?", n).collect()` (clippy::manual_repeat_n).
- crypto/init.rs (engine) — doc-lazy-continuation on a sentence
  following a bullet list; insert a blank line so it's a proper
  separate paragraph.
- engine_smoke test — three `B64.encode(&[N; 12])` had needless
  borrows; fixed to `B64.encode([N; 12])`.

Workspace `cargo clippy --all-targets --all-features -D warnings`
is now clean. Workspace tests still pass (no behavioural changes).

This unblocks the Ubuntu CI lane on PR #85.

+++ crates/assay-vault/src/dynamic.rs
@@ -408,7 +408,7 @@ pub mod aws_provider {
                 uuid::Uuid::new_v4().simple()
             );
             // STS minimum is 900s; clamp.
-            let duration = ttl_secs.max(900) as u64;
+            let duration = ttl_secs.max(900);
             let body = format!(
                 "Action=AssumeRole&Version=2011-06-15&RoleArn={}&RoleSessionName={}&DurationSeconds={}",
                 urlencode(&cfg.role_arn),
